Fixed the validation and sign in logic
