+++ lib/ts/recipe/dashboard/recipeImplementation.ts
@@ -13,8 +13,11 @@
  * under the License.
  */
 
+import NormalisedURLPath from "../../normalisedURLPath";
+import { Querier } from "../../querier";
 import { dashboardVersion } from "../../version";
 import { RecipeInterface } from "./types";
+import { validateApiKey } from "./utils";
 
 export default function getRecipeImplementation(): RecipeInterface {
     return {
@@ -22,21 +25,24 @@ export default function getRecipeImplementation(): RecipeInterface {
             return `https://cdn.jsdelivr.net/gh/supertokens/dashboard@v${dashboardVersion}/build/`;
         },
         shouldAllowAccess: async function (input) {
-            // For cases where we're not using the API key, the JWT is being used; we allow their access by default
-            if (!input.config.apiKey) {
-                return true;
-            }
-
-            let apiKeyHeaderValue: string | undefined = input.req.getHeaderValue("authorization");
-
-            // We receieve the api key as `Bearer API_KEY`, this retrieves just the key
-            apiKeyHeaderValue = apiKeyHeaderValue?.split(" ")[1];
-
-            if (apiKeyHeaderValue === undefined) {
+            try {
+                // For cases where we're not using the API key, the JWT is being used; we allow their access by default
+                if (!input.config.apiKey) {
+                    // make the check for the API endpoint here with querier
+                    let querier = Querier.getNewInstanceOrThrowError(undefined);
+                    const authHeaderValue = input.req.getHeaderValue("authorization")?.split(" ")[1];
+                    const sessionVerificationResponse = await querier.sendPostRequest(
+                        new NormalisedURLPath("/recipe/dashboard/session/verify"),
+                        {
+                            sessionId: authHeaderValue,
+                        }
+                    );
+                    return sessionVerificationResponse.status === "OK";
+                }
+                return await validateApiKey(input);
+            } catch (error) {
                 return false;
             }
-
-            return apiKeyHeaderValue === input.config.apiKey;
         },
     };
 }
